Mirgrated team form, added update functionality
